requirements.txt
Flask
python-dotenv
pandas
requests
langchain
openai
faiss-cpu
